[models.iosys] move two methods from reductors.h2
